engine/python/pyproject.toml
[project]
name = "watermark-engine"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "grpcio",
  "grpcio-tools",
  "protobuf",
  "onnxruntime",
  "opencv-python-headless",
]
